refactor(animation): split FrameAnimation out of Spring.swift
This change pushed Spring.swift to 331 lines — inside the 350
ceiling but past the point AGENTS.md §2.1 says to split, and one
addition away from crossing it. The file already held two types
with a clean seam: `Spring` is the scalar integrator and its
stability bound, `FrameAnimation` is the four-spring frame with
the retarget and re-seat surface. 127 and 208 lines apiece.

Pure move, no logic change. Re-pins the routing guard's entry
from Spring.swift to FrameAnimation.swift.
